Add basic Patmos files and configuration based on RISC-V.
This commit adds a copy of the RISC-V folder renamed to Patmos, and adds
configuration options to build CompCert for the Patmos architecture.

Since GCC does not exist for Patmos, it is not possible to actually build
the CompCert executable, but the proofs and extraction of the Coq code works.
Additionally, the runtime is simply a copy of the RISC-V runtime, which
obviously will not compile or run on a Patmos machine.

Some parts of the configuration may not be correct (especially the changes in
Machine.ml), and will probably have to be changed later on.